# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "FindReplaceRequest": {
        "description": "Literal (non-regex) replace across template sides and styling.",
        "properties": {
          "back": {
            "default": true,
            "title": "Back",
            "type": "boolean"
          },
          "css": {
            "default": true,
            "title": "Css",
            "type": "boolean"
          },
          "find": {
            "title": "Find",
            "type": "string"
          },
          "front": {
            "default": true,
            "title": "Front",
            "type": "boolean"
          },
          "modelName": {
            "title": "Modelname",
            "type": "string"
          },
          "replace": {
            "title": "Replace",
            "type": "string"
          }
        },
        "required": [
          "find",
          "replace"
        ],
        "title": "FindReplaceRequest",
        "type": "object"
      },
      "FindReplaceResult": {
        "description": "`affected` counts models that actually contained the text.",
        "properties": {
          "affected": {
            "title": "Affected",
            "type": "integer"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "affected",
          "stats"
        ],
        "title": "FindReplaceResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MutationResult_Any_": {
        "properties": {
          "result": {
            "title": "Result"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "stats"
        ],
        "title": "MutationResult[Any]",
        "type": "object"
      },
...
      "Paginated_Union_Any__ProjectedObject__str__int__float__bool__NoneType__": {
        "properties": {
          "items": {
            "items": {
              "anyOf": [
                {},
                {
                  "$ref": "#/components/schemas/ProjectedObject"
                },
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          },
          "stats": {
            "title": "Stats",
            "type": "object"
          }
        },
        "required": [
          "items",
          "stats"
        ],
        "title": "Paginated[Union[Any, ProjectedObject, str, int, float, bool, NoneType]]",
        "type": "object"
      },
...
      "ProjectedObject": {
        "properties": {},
        "title": "ProjectedObject",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Required when an API key is configured in Tsunagi settings.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Alternative to the X-API-Key header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT"
    },
    "title": "Tsunagi"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models": {
      "get": {
        "description": "Query models with filtering, field selection, and pagination. Note types define the structure of cards in Anki. Select note_count for the live number of notes using each type across all decks, or filter with where=note_count>0. Counts are included in full native query results; name/ID-only queries do not calculate them.",
        "operationId": "listModels",
        "parameters": [
          {
            "description": "Comma-separated fields to return",
            "in": "query",
            "name": "select",
            "required": false,
            "schema": {
              "description": "Comma-separated fields to return",
              "title": "Select",
              "type": "string"
            }
          },
          {
            "description": "Filter clauses (can specify multiple)",
            "in": "query",
            "name": "where",
            "required": false,
            "schema": {
              "description": "Filter clauses (can specify multiple)",
              "items": {
                "type": "string"
              },
              "title": "Where",
              "type": "array"
            }
          },
          {
            "description": "Anki search string (e.g. 'deck:Japanese tag:verb'). Only supported by search-backed resources; others return 400.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "description": "Anki search string (e.g. 'deck:Japanese tag:verb'). Only supported by search-backed resources; others return 400.",
              "title": "Search",
              "type": "string"
            }
          },
          {
            "description": "Response shape: auto, object, or scalar",
            "in": "query",
            "name": "shape",
            "required": false,
            "schema": {
              "default": "auto",
              "description": "Response shape: auto, object, or scalar",
              "title": "Shape",
              "type": "string"
            }
          },
          {
            "description": "Maximum results in this response. Omit to return all matches; no fixed upper cap.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "description": "Maximum results in this response. Omit to return all matches; no fixed upper cap.",
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque next_cursor from the previous response. Omit to start at page one; malformed or empty cursors return 400.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "description": "Opaque next_cursor from the previous response. Omit to start at page one; malformed or empty cursors return 400.",
              "title": "Cursor",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_Union_Any__ProjectedObject__str__int__float__bool__NoneType__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "List models",
        "tags": [
          "Models"
        ],
        "x-permission": "read:models"
      },
      "post": {
        "description": "Create a new model. Note types define the structure of cards in Anki. Select note_count for the live number of notes using each type across all decks, or filter with where=note_count>0. Counts are included in full native query results; name/ID-only queries do not calculate them.",
        "operationId": "createModel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "description": "Model data to create",
                "title": "Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult_Any_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create model",
        "tags": [
          "Models"
        ],
        "x-permission": "write:models"
      }
    },
...
    "/v1/models:find-replace": {
      "post": {
        "description": "Literal (not regex) replacement in card templates and CSS. Omit `model_name` to sweep every model. Only models that actually contained the text are touched or counted.",
        "operationId": "modelsFindReplace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindReplaceResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "BearerAuth": []
          }
        ],
        "summary": "Find and replace across templates and styling",
        "tags": [
          "Models"
        ],
        "x-permission": "write:notes"
      }
    },
...
  },
...
}